+++ mix/my_app_cldr.ex
@@ -1,6 +1,6 @@
 defmodule MyApp.Cldr do
   use Cldr,
     locales: ["en", "th"],
-    providers: [Cldr.Number, Cldr.Unit, Cldr.Territory]
+    providers: [Cldr.Number, Cldr.Unit, Cldr.Territory, Cldr.LocaleDisplay]
 
 end